Extract wearables settings panel
